Agregando validacion si existe o no usuario
