+++ src/Routes.js
@@ -18,72 +18,72 @@ function Routes() {
     return <Redirect to="/" />;
   };
 
-  const token = window.localStorage.getItem("token");
-  const [user, setUser] = useState({});
-  let idUser = undefined;
-  console.log("soy user11111-----routes>", user);
+  // const token = window.localStorage.getItem("token");
+  // const [user, setUser] = useState({});
+  // let idUser = undefined;
+  // console.log("soy user11111-----routes>", user);
 
-  const obtenerDatos = () => {
-    console.log("entrando 2222");
-    if (token) {
-      console.log("entrando");
-      const user2 = payload();
-      idUser = user2.id;
-      const config = {
-        headers: {
-          Authorization: `JWT ${token}`,
-        },
-      };
+  // const obtenerDatos = () => {
+  //   console.log("entrando 2222");
+  //   if (token != undefined) {
+  //     console.log("entrando");
+  //     const user2 = payload();
+  //     idUser = user2.id;
+  //     const config = {
+  //       headers: {
+  //         Authorization: `JWT ${token}`,
+  //       },
+  //     };
 
-      const obtenerUser = async () => {
-        await axios
-          .get(
-            `https://ecomerce-master.herokuapp.com/api/v1/user/${idUser}`,
-            config
-          )
-          .then((res) => {
-            console.log("obteniendo data de usuario", res.data, res.status);
-            setUser(res.data);
-            console.log("soy user2222-----routes>", user);
-          })
+  //     const obtenerUser = async () => {
+  //       await axios
+  //         .get(
+  //           `https://ecomerce-master.herokuapp.com/api/v1/user/${idUser}`,
+  //           config
+  //         )
+  //         .then((res) => {
+  //           console.log("obteniendo data de usuario", res.data, res.status);
+  //           setUser(res.data);
+  //           console.log("soy user2222-----routes>", user);
+  //         })
 
-          .catch((error) => {
-            console.error(error.response.data);
-          });
-      };
-      obtenerUser();
-      console.log("soy user aslkdjasdlasjdlaks", user);
-      console.log("soy el user activo", user.first_name);
-    }
-  };
+  //         .catch((error) => {
+  //           console.error(error.response.data);
+  //         });
+  //     };
+  //     obtenerUser();
+  //     console.log("soy user aslkdjasdlasjdlaks", user);
+  //     console.log("soy el user activo", user.first_name);
+  //   }
+  // };
 
-  useEffect(() => {
-    // console.log("itemSearch   ", itemsSearch);
-    // obtenerUser();
-    obtenerDatos();
-    console.log("sourle,mlmlmlmlmlkdalskdmalskdlaskjlkasjflaksjfl", user);
-  }, [token]);
+  // useEffect(() => {
+  //   // console.log("itemSearch   ", itemsSearch);
+  //   // obtenerUser();
+  //   obtenerDatos();
+  //   console.log("sourle,mlmlmlmlmlkdalskdmalskdlaskjlkasjflaksjfl", user);
+  // }, [token]);
 
   return (
     <>
       <Route exact path="/">
         {/* <Home /> */}
-        <ProductsList user={user} />
+        <ProductsList />
       </Route>
       <Route exact path="/search/:itemsSearch">
-        <ProductsList user={user} />
+        <ProductsList />
       </Route>
-      <Route exact path="/login" user={user}>
+      <Route exact path="/login">
         <Login />
       </Route>
       <Route exact path="/signup">
         <Signup />
       </Route>
       <Route exact path="/item/:iditem">
-        <SingleProduct user={user} />
+        <SingleProduct />
       </Route>
       <Route path="/profile">
-        <Perfil user={user} />
+        <Perfil />
       </Route>
       <Route path="/create-product">
         <CreateProduct />
